Changing temporary directory to /bot-tmp
Builder copies files in /tmp to (e.g. `/tmp/highstate.sh`) to execute
them on a target EC2 instance.

However, when a Salt state mounts something over /tmp the process fails
at the end when trying to remove `/tmp/highstate.sh`, as the original
/tmp folder has been shadowed by the new mount.

Therefore, in the arm wrestling between bot and builder, builder gets to
use /tmp.

This change does not affect any production instance, as I would like to
get it in before (re)creating `elife-bot--prod` and porting to it.