Running cleanup cron every hour
This should not interfere with running workflows as it only acts on files older than 2 days.

When instances are turned on, they stay up for at least an hour, so this should suffice to clean up their /bot-tmp folder before they are shutdown again. Only executing it at 2 AM sometimes lead to the cleanup not being executed for several days until the disk fills up.